fix(lint): gosec: G115 error

+++ cp.go
@@ -210,7 +210,8 @@ func untar(reader io.Reader, dst, prefix string) error {
 				return err
 			}
 		case tar.TypeReg:
-			if err = createFile(tarReader, dstPath, os.FileMode(uint32(header.Mode))); err != nil {
+			//nolint:gosec
+			if err = createFile(tarReader, dstPath, os.FileMode(header.Mode)); err != nil {
 				return err
 			}
 		}
